Refactor Counter class and update tests accordingly
The Counter class has been refactored to a class with static methods, adding the ability to accept a $shellAccess parameter. Consequently, the console.php file was deprecated and several modifications were made to the unit tests to match the updated functionality of the Counter class. Minor changes were also made to some classes implementing the CounterInterface.